Add SinglyLinkedList class for Dijkstra's algorithm implementation.

+++ DirectedGraph/Edge.cs
@@ -7,7 +7,7 @@ namespace DirectedGraph
         TEdgeId Id { get; }
         INode<TNodeId, TEdgeId> FromNode { get; }
         INode<TNodeId, TEdgeId> ToNode { get; }
-        int Weight { get; }
+        double Weight { get; }
     }
 
     class Edge<TNodeId, TEdgeId, TData> : IEdge<TNodeId, TEdgeId>
@@ -15,9 +15,9 @@ namespace DirectedGraph
         public TEdgeId Id { get; }
         public INode<TNodeId, TEdgeId> FromNode { get; }
         public INode<TNodeId, TEdgeId> ToNode { get; }
-        public int Weight { get; private set; }
+        public double Weight { get; private set; }
 
-        public Edge(TEdgeId id, INode<TNodeId, TEdgeId> from, INode<TNodeId, TEdgeId> to, int weight)
+        public Edge(TEdgeId id, INode<TNodeId, TEdgeId> from, INode<TNodeId, TEdgeId> to, double weight)
         {
             Id = id;
             FromNode = from;
